+++ AituEmbedded/WebBridge.swift
@@ -5,36 +5,7 @@ protocol WebBridgeController {
     func receive(body: Any, from url: URL?, sender: WebBridgeSender)
 }
 
-protocol WebBridgeSender {
-    func send(reply: String)
-}
-
-final class WebBridge {
-    struct Registrator {
-        private let _register: (WKScriptMessageHandler, String) -> Void
-
-        init(register: @escaping (WKScriptMessageHandler, String) -> Void) {
-            _register = register
-        }
-
-        func register(_ receiver: WKScriptMessageHandler, method: String) {
-            _register(receiver, method)
-        }
-    }
-    private let receivers: [MessageReceiver]
-
-    init(controllers: [WebBridgeController], registrator: Registrator, sender: WebBridgeSender) {
-        receivers = controllers.map({ contoller in
-            let receiver = MessageReceiver(receive: { body, url in
-                contoller.receive(body: body, from: url, sender: sender)
-            })
-            registrator.register(receiver, method: contoller.method)
-            return receiver
-        })
-    }
-}
-
-private final class MessageReceiver: NSObject, WKScriptMessageHandler {
+final class MessageReceiver: NSObject, WKScriptMessageHandler {
     private let receive: (Any, URL?) -> Void
 
     init(receive: @escaping (Any, URL?) -> Void) {
@@ -47,17 +18,41 @@ private final class MessageReceiver: NSObject, WKScriptMessageHandler {
     }
 }
 
-extension WKWebView: WebBridgeSender {
-    func send(reply js: String) {
+public protocol WebBridgeSender {
+    func send(reply: String)
+}
+
+public protocol WebBridgeRegistrator {
+    func register(_ receiver: WKScriptMessageHandler, method: String)
+}
+
+public struct AituWebBridgeFormatter: WebBridgeSender {
+    private let sender: (String) -> Void
+
+    public init(sender: @escaping (String) -> Void) {
+        self.sender = sender
+    }
+
+    public func send(reply: String) {
+        let payloadPlace = "###payload###"
+        let template = """
+        window.dispatchEvent(new CustomEvent("aituEvents", {
+            detail: \(payloadPlace)
+        }))
+        """
+        let script = template.replacingOccurrences(of: payloadPlace, with: reply)
+        sender(script)
+    }
+}
+
+extension WKWebView: WebBridgeSender, WebBridgeRegistrator {
+    public func send(reply: String) {
         DispatchQueue.main.async {
-            let payloadPlace = "###payload###"
-            let template = """
-            window.dispatchEvent(new CustomEvent("aituEvents", {
-                detail: \(payloadPlace)
-            }))
-            """
-            let script = template.replacingOccurrences(of: payloadPlace, with: js)
-            self.evaluateJavaScript(script, completionHandler: nil)
+            self.evaluateJavaScript(reply, completionHandler: nil)
         }
+    }
+
+    public func register(_ receiver: WKScriptMessageHandler, method: String) {
+        configuration.userContentController.add(receiver, name: method)
     }
 }
